pyproject.toml
[tool.poetry]
name = "celsiusbot"
version = "0.1.0"
description = ""
authors = ["Aleksandr Zuev <zuev08@gmail.com>"]
readme = "README.md"
packages = [{include = "celsiusbot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.86.0"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
aiogram = "^2.23.1"
python-socketio = "^5.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
